fix(test): release memory search fixtures after failed phases
